feat: add mappers for Electrical DTOs and update ID type
This commit introduces mapper functions to convert between Electrical BAP/Report DTOs and their corresponding Domain models. It also updates the `extraId` field type from `Int` to `Long` in `ElectricalBapDto.kt` and `ElectricalReportDto.kt`.

Key changes:
- `app/src/main/java/com/nakersolutionid/nakersolutionid/data/remote/dto/electrical/ElectricalBapDto.kt`:
    - Changed `extraId` type from `Int` to `Long` in `ElectricalBapReportData` and `ElectricalBapRequest`.
- `app/src/main/java/com/nakersolutionid/nakersolutionid/data/remote/dto/electrical/ElectricalReportDto.kt`:
    - Changed `extraId` type from `Int` to `Long` in `ElectricalReportRequest` and `ElectricalReportData`.
- `app/src/main/java/com/nakersolutionid/nakersolutionid/data/remote/mapper/ElectricalDtoMapper.kt`:
    - Added new file with mapper functions:
        - `InspectionWithDetailsDomain.toElectricalBapRequest()`: Converts domain model to BAP request DTO.
        - `ElectricalBapReportData.toInspectionWithDetailsDomain()`: Converts BAP report DTO to domain model.
        - `InspectionWithDetailsDomain.toElectricalReportRequest()`: Converts domain model to Report request DTO.
        - `ElectricalReportData.toInspectionWithDetailsDomain()`: Converts Report DTO to domain model.
    - These mappers handle the transformation of nested data structures for general data, technical data, visual inspections, testing results, document checks, and findings/recommendations.
    - Defined internal category constants for consistent mapping between UI, Domain, and DTO layers.

+++ app/src/main/java/com/nakersolutionid/nakersolutionid/data/remote/dto/electrical/ElectricalReportDto.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class ElectricalReportRequest(
     @SerializedName("examinationType") val examinationType: String,
     @SerializedName("inspectionType") val inspectionType: String,
-    @SerializedName("extraid") val extraId: Int,
+    @SerializedName("extraid") val extraId: Long,
     @SerializedName("equipmentType") val equipmentType: String,
     @SerializedName("createdaAt") val createdAt: String,
     @SerializedName("generalData") val generalData: ElectricalGeneralData,
@@ -35,7 +35,7 @@ data class ElectricalReportData(
     @SerializedName("id") val id: String, // From response JSON
     @SerializedName("examinationType") val examinationType: String,
     @SerializedName("inspectionType") val inspectionType: String,
-    @SerializedName("extraid") val extraId: Int,
+    @SerializedName("extraid") val extraId: Long,
     @SerializedName("createdaAt") val createdAt: String, // Note: Typo in JSON "createdaAt"
     @SerializedName("equipmentType") val equipmentType: String,
     @SerializedName("generalData") val generalData: ElectricalGeneralData,
